requirements.txt
ipykernel==6.29.4
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
